feat/cli Renamed `src/SrcCommand.php` to `src/Src/SrcCommand.php`.
