+++ src/Src/SrcCommand.php
@@ -1,0 +1,24 @@
+<?php
+
+namespace Zerotoprod\Psr4ClassnameCli\Src;
+
+use Symfony\Component\Console\Attribute\AsCommand;
+use Symfony\Component\Console\Command\Command;
+use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Output\OutputInterface;
+
+#[AsCommand(
+    name: SrcCommand::signature,
+    description: 'Project source link'
+)]
+class SrcCommand extends Command
+{
+    public const signature = 'psr4-classname-cli:src';
+
+    protected function execute(InputInterface $input, OutputInterface $output): int
+    {
+        $output->writeln('https://github.com/zero-to-prod/psr4-classname-cli');
+
+        return Command::SUCCESS;
+    }
+}
